refactor: dedupe socket and logging helpers
