+++ src/main.rs
@@ -28,7 +28,9 @@ mod events;
 mod ghostty;
 mod input;
 mod integration;
+mod ipc;
 mod layout;
+mod logging;
 mod pane;
 mod persist;
 mod platform;
@@ -43,34 +45,7 @@ mod update;
 mod workspace;
 
 fn init_logging() {
-    use std::fs::{self, OpenOptions};
-    use tracing_subscriber::EnvFilter;
-
-    let log_dir = crate::config::config_dir();
-    let _ = fs::create_dir_all(&log_dir);
-    let log_path = log_dir.join("herdr.log");
-
-    // Rotate: truncate if over 5MB
-    if let Ok(meta) = fs::metadata(&log_path) {
-        if meta.len() > 5 * 1024 * 1024 {
-            let _ = fs::remove_file(&log_path);
-        }
-    }
-
-    let file = match OpenOptions::new().create(true).append(true).open(&log_path) {
-        Ok(f) => f,
-        Err(_) => return, // can't open log file, proceed without logging
-    };
-
-    let filter =
-        EnvFilter::try_from_env("HERDR_LOG").unwrap_or_else(|_| EnvFilter::new("herdr=info"));
-
-    let _ = tracing_subscriber::fmt()
-        .with_env_filter(filter)
-        .with_writer(file)
-        .with_ansi(false)
-        .with_target(false)
-        .try_init();
+    crate::logging::init_file_logging("herdr.log");
 }
 
 const DEFAULT_CONFIG: &str = r##"# herdr configuration
